api/account/factories.py
import factory
from factory.faker import faker
from .models import User

faker = faker.Faker()


class UserFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = User

    username = factory.Sequence(lambda n: f"user{n}")
    email = factory.Sequence(lambda n: f"user{n}@gmail.com")
    password = factory.django.Password("parol123")
