requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==20.1.0
kombu==5.3.4
mysqlclient==2.2.0
prompt-toolkit==3.0.41
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
